app-history/per-entry-events/beforedispose-same-document-with-transitionWhile.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async (t) => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(r => window.onload = () => t.step_timeout(r, 0));

  location.hash = "#1";
  location.hash = "#2";
  location.hash = "#3";

  assert_equals(appHistory.entries().length, 4);
  const [entry0, entry1, entry2, entry3] = appHistory.entries();
  assert_equals((new URL(entry2.url)).hash, "#2");
  assert_equals((new URL(entry3.url)).hash, "#3");

  let beforedispose2Called = false;
  entry2.onbeforedispose = t.step_func(e => {
    beforedispose2Called = true;

    assert_equals(e.constructor, Event);
    assert_equals(e.bubbles, false);
    assert_equals(e.cancelable, false);
    assert_equals(e.composed, false);

    assert_array_equals(
      appHistory.entries(),
      [entry0, entry1, entry2, entry3],
      "during beforedispose for entry 2 the entries must not have been modified");
    assert_equals(location.hash, "#1", "location.hash still on #1 during beforedispose for entry 2");
    assert_equals(appHistory.current, entry1, "current entry is still entry 1 during beforedispose for entry 2");
    assert_equals(appHistory.transition.from, entry1, "transition from during beforedispose for entry 3");
    assert_equals(appHistory.transition.navigationType, "push", "transition navigationType during beforedispose for entry 3");
    assert_true(appHistory.canGoBack, "canGoBack is still true during beforedispose for entry 2");
    assert_true(appHistory.canGoForward, "canGoForward is still true during beforedispose for entry 2");
  });

  let beforedispose3Called = false;
  entry3.addEventListener("beforedispose", t.step_func(e => {
    beforedispose3Called = true;

    assert_true(beforedispose2Called, "beforedispose for entry 2 must have happened before entry 3");
    assert_array_equals(
      appHistory.entries(),
      [entry0, entry1, entry2, entry3],
      "during beforedispose for entry 3 the entries must not have been modified");
    assert_equals(location.hash, "#1", "location.hash still on #1 during beforedispose for entry 3");
    assert_equals(appHistory.current, entry1, "current entry is still entry 1 during beforedispose for entry 3");
    assert_equals(appHistory.transition.from, entry1, "transition from during beforedispose for entry 3");
    assert_equals(appHistory.transition.navigationType, "push", "transition navigationType during beforedispose for entry 3");
    assert_true(appHistory.canGoBack, "canGoBack is still true during beforedispose for entry 3");
    assert_true(appHistory.canGoForward, "canGoForward is still true during beforedispose for entry 3");
  }));

  await appHistory.goTo(entry1.key);

  appHistory.addEventListener("navigate", e => {
    e.transitionWhile(Promise.resolve());
  });

  await appHistory.navigate("#fork");

  assert_true(beforedispose3Called, "beforedispose for entry 3 must have happened");

  assert_equals(appHistory.entries().length, 3);
  const [finalEntry0, finalEntry1, finalEntry2] = appHistory.entries();
  assert_equals(finalEntry0, entry0);
  assert_equals(finalEntry1, entry1);
  assert_not_equals(finalEntry2, entry2);
  assert_equals(appHistory.current, finalEntry2);
  assert_equals((new URL(finalEntry2.url)).hash, "#fork");
}, "beforedispose events when disposing same-document entries");
</script>
